feat: Create DTO as middle layer between Repositories and Services

+++ Modules/UI/routes/api.php
@@ -11,15 +11,36 @@ Route::prefix('v1')->group(function () {
     |--------------------------------------------------------------------------
     |
     */
+
     Route::prefix('auth')->controller(AuthController::class)->group(function () {
         Route::post('login', 'login');
         Route::post('register', 'register');
-        Route::post('{id}/verify-otp/{otp}', 'verifyOtp')->whereUlid('id');
-        Route::post('refresh', 'refresh')->middleware('auth-user');
-        Route::post('logout', 'logout')->middleware('auth-user');
+        Route::post('{user}/verify-otp/{otp}', 'verifyOtp')->whereUlid('user');
         Route::post('forgot-password', 'forgotPassword');
-        Route::post('reset-password/{token}', 'resetPassword')->middleware('guest')->name('password.reset');
-        Route::get('me', 'me')->middleware('auth-user');
+        Route::post('reset-password/{token}', 'resetPassword')->name('password.reset');
+        Route::middleware('auth-user')->group(function () {
+            Route::post('refresh', 'refresh');
+            Route::post('logout', 'logout');
+            Route::get('me', 'me');
+            Route::post('{user}/upload-image', 'storeProfileImage')->whereUlid('user');
+        });
+    });
+
+    /*
+    |--------------------------------------------------------------------------
+    |  Admin
+    |--------------------------------------------------------------------------
+    |
+    */
+
+    Route::prefix('admin')->group(function () {
+        Route::prefix('students')->middleware(['auth-user', 'role:super_admin,admin'])->group(function () {
+            Route::get('/', fn () => 'hello-students')->middleware('permission:view_students');
+        });
+
+        Route::prefix('teachers')->middleware(['auth-user', 'role:super_admin,admin'])->group(function () {
+            Route::get('/', fn () => 'hello-teachers')->middleware('permission:view_teachers');
+        });
     });
 
     /*
@@ -28,13 +49,14 @@ Route::prefix('v1')->group(function () {
     |--------------------------------------------------------------------------
     |
     */
-    Route::middleware(['auth-user', 'role:super_admin'])->prefix('settings')->controller(SettingController::class)->group(function () {
-        Route::get('/', 'index');
-        Route::get('/{id}', 'show');
-        Route::post('/', 'store');
-        Route::put('/{id}', 'update');
-        Route::delete('/{id}', 'destroy');
-        Route::patch('/{id}', 'restore');
-        Route::delete('/{id}/force', 'forceDelete');
+
+    Route::middleware(['auth-user', 'role:super_admin,admin'])->prefix('settings')->controller(SettingController::class)->group(function () {
+        Route::get('/', 'index')->middleware('permission:view_settings');
+        Route::get('/{id}', 'show')->middleware('permission:view_settings');
+        Route::post('/', 'store')->middleware('permission:create_settings');
+        Route::put('/{id}', 'update')->middleware('permission:update_settings');
+        Route::delete('/{id}', 'destroy')->middleware('permission:delete_settings');
+        Route::patch('/{id}', 'restore')->middleware('permission:restore_settings');
+        Route::delete('/{id}/force', 'forceDelete')->middleware('permission:force_delete_settings');
     });
 });
